+++ pkg/framesql/frame.go
@@ -23,7 +23,8 @@ func EvaluateInFrame(expression string, input *data.Frame) (interface{}, error) 
 	}
 	if input != nil {
 		for _, field := range input.Fields {
-			parameters[slugifyFieldName(field.Name)] = field
+			parameters[SlugifyFieldName(field.Name)] = field
+			parameters[field.Name] = field
 		}
 	}
 	result, err := parsedExpression.Evaluate(parameters)
@@ -35,6 +36,9 @@ func EvaluateInFrame(expression string, input *data.Frame) (interface{}, error) 
 
 var expressionFunctions = map[string]govaluate.ExpressionFunction{
 	"count": func(arguments ...interface{}) (interface{}, error) {
+		if len(arguments) < 1 {
+			return nil, errors.New("invalid arguments to count method")
+		}
 		field, ok := arguments[0].(*data.Field)
 		if !ok {
 			return nil, errors.New("first argument is not a valid field")
@@ -42,6 +46,9 @@ var expressionFunctions = map[string]govaluate.ExpressionFunction{
 		return float64(field.Len()), nil
 	},
 	"first": func(arguments ...interface{}) (interface{}, error) {
+		if len(arguments) < 1 {
+			return nil, errors.New("invalid arguments to first method")
+		}
 		field, ok := arguments[0].(*data.Field)
 		if !ok {
 			return nil, errors.New("first argument is not a valid field")
@@ -49,9 +56,21 @@ var expressionFunctions = map[string]govaluate.ExpressionFunction{
 		if v, ok := field.At(0).(*float64); ok {
 			return *v, nil
 		}
-		return nil, errors.New("unable to find first value")
+		if v, ok := field.At(0).(*int); ok {
+			return *v, nil
+		}
+		if v, ok := field.At(0).(*string); ok {
+			return *v, nil
+		}
+		if v, ok := field.At(0).(*bool); ok {
+			return *v, nil
+		}
+		return field.At(0), nil
 	},
 	"last": func(arguments ...interface{}) (interface{}, error) {
+		if len(arguments) < 1 {
+			return nil, errors.New("invalid arguments to last method")
+		}
 		field, ok := arguments[0].(*data.Field)
 		if !ok {
 			return nil, errors.New("first argument is not a valid field")
@@ -59,16 +78,28 @@ var expressionFunctions = map[string]govaluate.ExpressionFunction{
 		if v, ok := field.At(field.Len() - 1).(*float64); ok {
 			return *v, nil
 		}
+		if v, ok := field.At(field.Len() - 1).(*int); ok {
+			return *v, nil
+		}
+		if v, ok := field.At(field.Len() - 1).(*string); ok {
+			return *v, nil
+		}
+		if v, ok := field.At(field.Len() - 1).(*bool); ok {
+			return *v, nil
+		}
 		return nil, errors.New("unable to find first value")
 	},
 	"sum": func(arguments ...interface{}) (interface{}, error) {
+		if len(arguments) < 1 {
+			return nil, errors.New("invalid arguments to sum method")
+		}
 		field, ok := arguments[0].(*data.Field)
 		if !ok {
 			return nil, errors.New("first argument is not a valid field")
 		}
 		sum := float64(0)
 		for i := 0; i < field.Len(); i++ {
-			if v, ok := field.At(i).(*float64); ok {
+			if v, ok := toFloat64p(field.At(i)); ok {
 				sum += *v
 			}
 
@@ -76,13 +107,16 @@ var expressionFunctions = map[string]govaluate.ExpressionFunction{
 		return sum, nil
 	},
 	"min": func(arguments ...interface{}) (interface{}, error) {
+		if len(arguments) < 1 {
+			return nil, errors.New("invalid arguments to min method")
+		}
 		field, ok := arguments[0].(*data.Field)
 		if !ok {
 			return nil, errors.New("first argument is not a valid field")
 		}
 		var min *float64
 		for i := 0; i < field.Len(); i++ {
-			if v, ok := field.At(i).(*float64); ok {
+			if v, ok := toFloat64p(field.At(i)); ok {
 				if min == nil || *v < *min {
 					min = v
 				}
@@ -91,13 +125,16 @@ var expressionFunctions = map[string]govaluate.ExpressionFunction{
 		return *min, nil
 	},
 	"max": func(arguments ...interface{}) (interface{}, error) {
+		if len(arguments) < 1 {
+			return nil, errors.New("invalid arguments to max method")
+		}
 		field, ok := arguments[0].(*data.Field)
 		if !ok {
 			return nil, errors.New("first argument is not a valid field")
 		}
 		var max *float64
 		for i := 0; i < field.Len(); i++ {
-			if v, ok := field.At(i).(*float64); ok {
+			if v, ok := toFloat64p(field.At(i)); ok {
 				if max == nil || *v > *max {
 					max = v
 				}
@@ -106,13 +143,16 @@ var expressionFunctions = map[string]govaluate.ExpressionFunction{
 		return *max, nil
 	},
 	"mean": func(arguments ...interface{}) (interface{}, error) {
+		if len(arguments) < 1 {
+			return nil, errors.New("invalid arguments to mean method")
+		}
 		field, ok := arguments[0].(*data.Field)
 		if !ok {
 			return nil, errors.New("first argument is not a valid field")
 		}
 		sum := float64(0)
 		for i := 0; i < field.Len(); i++ {
-			if v, ok := field.At(i).(*float64); ok {
+			if v, ok := toFloat64p(field.At(i)); ok {
 				sum += *v
 			}
 
@@ -124,8 +164,50 @@ var expressionFunctions = map[string]govaluate.ExpressionFunction{
 	},
 }
 
-func slugifyFieldName(input string) string {
+func SlugifyFieldName(input string) string {
 	re, _ := regexp.Compile(`[^\w]`)
 	input = strings.TrimSpace(re.ReplaceAllString(strings.ToLower(strings.TrimSpace(input)), "_"))
 	return input
+}
+
+func toFloat64p(input interface{}) (*float64, bool) {
+	if v, ok := input.(*float64); ok {
+		return v, true
+	}
+	if v, ok := input.(float64); ok {
+		return &v, true
+	}
+	if v, ok := input.(*float32); ok {
+		v1 := float64(*v)
+		return &v1, true
+	}
+	if v, ok := input.(float32); ok {
+		v1 := float64(v)
+		return &v1, true
+	}
+	if v, ok := input.(*int); ok {
+		v1 := float64(*v)
+		return &v1, true
+	}
+	if v, ok := input.(int); ok {
+		v1 := float64(v)
+		return &v1, true
+	}
+	if v, ok := input.(*int64); ok {
+		v1 := float64(*v)
+		return &v1, true
+	}
+	if v, ok := input.(int64); ok {
+		v1 := float64(v)
+		return &v1, true
+	}
+	if v, ok := input.(*int32); ok {
+		v1 := float64(*v)
+		return &v1, true
+	}
+	if v, ok := input.(int32); ok {
+		v1 := float64(v)
+		return &v1, true
+	}
+	return nil, false
 }