+++ pkg/infinity/csvBackend.go
@@ -8,7 +8,6 @@ import (
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/yesoreyeram/grafana-framer/csvFramer"
 	"github.com/yesoreyeram/grafana-framer/gframer"
-	"github.com/yesoreyeram/grafana-infinity-datasource/pkg/framesql"
 	querySrv "github.com/yesoreyeram/grafana-infinity-datasource/pkg/query"
 )
 
@@ -57,30 +56,20 @@ func GetCSVBackendResponse(responseString string, query querySrv.Query) (*data.F
 	if newFrame != nil {
 		frame.Fields = append(frame.Fields, newFrame.Fields...)
 	}
+	frame, err = GetFrameWithComputedColumns(frame, query.ComputedColumns)
+	if err != nil {
+		backend.Logger.Error("error getting computed column", "error", err.Error())
+		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
+		return frame, err
+	}
+	frame, err = ApplyFilter(frame, query.FilterExpression)
+	if err != nil {
+		backend.Logger.Error("error applying filter", "error", err.Error())
+		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
+		return frame, fmt.Errorf("error applying filter. %w", err)
+	}
 	if strings.TrimSpace(query.SummarizeExpression) != "" {
-		summary, err := framesql.EvaluateInFrame(query.SummarizeExpression, frame)
-		if err != nil {
-			backend.Logger.Error("error evaluating summarize expression", "error", err.Error())
-			frame.Meta.Custom = &CustomMeta{
-				Data:  responseString,
-				Query: query,
-				Error: err.Error(),
-			}
-			return frame, err
-		}
-		switch t := summary.(type) {
-		case float64:
-			v := summary.(float64)
-			summaryFrame := &data.Frame{
-				Name:   frame.Name,
-				RefID:  frame.RefID,
-				Fields: []*data.Field{data.NewField("summary", nil, []*float64{&v})},
-			}
-			summaryFrame.SetMeta(frame.Meta)
-			return summaryFrame, nil
-		default:
-			fmt.Print(t)
-		}
+		return GetSummaryFrame(frame, query.SummarizeExpression)
 	}
 	if query.Format == "timeseries" && frame.TimeSeriesSchema().Type == data.TimeSeriesTypeLong {
 		if wFrame, err := data.LongToWide(frame, &data.FillMissing{Mode: data.FillModeNull}); err == nil {