+++ pkg/infinity/inline.go
@@ -5,10 +5,8 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/yesoreyeram/grafana-framer/jsonFramer"
-	"github.com/yesoreyeram/grafana-infinity-datasource/pkg/framesql"
 	querySrv "github.com/yesoreyeram/grafana-infinity-datasource/pkg/query"
 )
 
@@ -36,39 +34,21 @@ func GetFrameForInlineSources(query querySrv.Query) (*data.Frame, error) {
 			Columns:      columns,
 		})
 		if err != nil {
-			backend.Logger.Error("error building frame", "error", err.Error())
-			frame.Meta.Custom = &CustomMeta{
-				Query: query,
-				Error: err.Error(),
-			}
 			return frame, err
 		}
 		if newFrame != nil {
 			frame.Fields = append(frame.Fields, newFrame.Fields...)
 		}
+		frame, err = GetFrameWithComputedColumns(frame, query.ComputedColumns)
+		if err != nil {
+			return frame, fmt.Errorf("error getting computed column. %w", err)
+		}
+		frame, err = ApplyFilter(frame, query.FilterExpression)
+		if err != nil {
+			return frame, fmt.Errorf("error applying filter. %w", err)
+		}
 		if strings.TrimSpace(query.SummarizeExpression) != "" {
-			summary, err := framesql.EvaluateInFrame(query.SummarizeExpression, frame)
-			if err != nil {
-				backend.Logger.Error("error evaluating summarize expression", "error", err.Error())
-				frame.Meta.Custom = &CustomMeta{
-					Query: query,
-					Error: err.Error(),
-				}
-				return frame, err
-			}
-			switch t := summary.(type) {
-			case float64:
-				v := summary.(float64)
-				summaryFrame := &data.Frame{
-					Name:   frame.Name,
-					RefID:  frame.RefID,
-					Fields: []*data.Field{data.NewField("summary", nil, []*float64{&v})},
-				}
-				summaryFrame.SetMeta(frame.Meta)
-				return summaryFrame, nil
-			default:
-				fmt.Print(t)
-			}
+			return GetSummaryFrame(frame, query.SummarizeExpression)
 		}
 		if query.Format == "timeseries" && frame.TimeSeriesSchema().Type == data.TimeSeriesTypeLong {
 			if wFrame, err := data.LongToWide(frame, &data.FillMissing{Mode: data.FillModeNull}); err == nil {