+++ pkg/infinity/jsonBackend.go
@@ -8,7 +8,6 @@ import (
 	"github.com/grafana/grafana-plugin-sdk-go/backend"
 	"github.com/grafana/grafana-plugin-sdk-go/data"
 	"github.com/yesoreyeram/grafana-framer/jsonFramer"
-	"github.com/yesoreyeram/grafana-infinity-datasource/pkg/framesql"
 	querySrv "github.com/yesoreyeram/grafana-infinity-datasource/pkg/query"
 )
 
@@ -17,6 +16,7 @@ func GetJSONBackendResponse(urlResponseObject interface{}, query querySrv.Query)
 	responseString, err := json.Marshal(urlResponseObject)
 	if err != nil {
 		backend.Logger.Error("error json parsing root data", "error", err.Error())
+		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
 		return frame, fmt.Errorf("error parsing json root data")
 	}
 	columns := []jsonFramer.ColumnSelector{}
@@ -33,46 +33,29 @@ func GetJSONBackendResponse(urlResponseObject interface{}, query querySrv.Query)
 		RootSelector: query.RootSelector,
 		Columns:      columns,
 	})
-	frame.Meta = &data.FrameMeta{
-		Custom: &CustomMeta{
-			Query: query,
-		},
-	}
+	frame.Meta = &data.FrameMeta{Custom: &CustomMeta{Query: query}}
 	if err != nil {
 		backend.Logger.Error("error getting response for query", "error", err.Error())
-		frame.Meta.Custom = &CustomMeta{
-			Query: query,
-			Error: err.Error(),
-		}
+		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
 		return frame, err
 	}
 	if newFrame != nil {
 		frame.Fields = append(frame.Fields, newFrame.Fields...)
 	}
+	frame, err = GetFrameWithComputedColumns(frame, query.ComputedColumns)
+	if err != nil {
+		backend.Logger.Error("error getting computed column", "error", err.Error())
+		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
+		return frame, err
+	}
+	frame, err = ApplyFilter(frame, query.FilterExpression)
+	if err != nil {
+		backend.Logger.Error("error applying filter", "error", err.Error())
+		frame.Meta.Custom = &CustomMeta{Query: query, Error: err.Error()}
+		return frame, fmt.Errorf("error applying filter. %w", err)
+	}
 	if strings.TrimSpace(query.SummarizeExpression) != "" {
-		summary, err := framesql.EvaluateInFrame(query.SummarizeExpression, frame)
-		if err != nil {
-			backend.Logger.Error("error evaluating summarize expression", "error", err.Error())
-			frame.Meta.Custom = &CustomMeta{
-				Data:  urlResponseObject,
-				Query: query,
-				Error: err.Error(),
-			}
-			return frame, err
-		}
-		switch t := summary.(type) {
-		case float64:
-			v := summary.(float64)
-			summaryFrame := &data.Frame{
-				Name:   frame.Name,
-				RefID:  frame.RefID,
-				Fields: []*data.Field{data.NewField("summary", nil, []*float64{&v})},
-			}
-			summaryFrame.SetMeta(frame.Meta)
-			return summaryFrame, nil
-		default:
-			fmt.Print(t)
-		}
+		return GetSummaryFrame(frame, query.SummarizeExpression)
 	}
 	if query.Format == "timeseries" && frame.TimeSeriesSchema().Type == data.TimeSeriesTypeLong {
 		if wFrame, err := data.LongToWide(frame, &data.FillMissing{Mode: data.FillModeNull}); err == nil {